task: created innovation signal store context
