+++ src/modules/shared/resolvers/innovation-data.resolver.ts
@@ -4,8 +4,7 @@ import { NGXLogger } from 'ngx-logger';
 import { Observable, of } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
 
-import { AuthenticationStore } from '@modules/stores/authentication/authentication.store';
-import { ContextStore } from '@modules/stores/context/context.store';
+import { AuthenticationStore, InnovationContextStore } from '@modules/stores';
 
 /**
  * Note: With the creation of the context store, this can be changed to a guard in the future,
@@ -17,20 +16,17 @@ export class InnovationDataResolver {
     private router: Router,
     private logger: NGXLogger,
     private authenticationStore: AuthenticationStore,
-    private contextStore: ContextStore
+    private innovationStoreService: InnovationContextStore
   ) {}
 
   resolve(route: ActivatedRouteSnapshot): Observable<null | { id: string; name: string }> {
-    return this.contextStore
+    return this.innovationStoreService
       .getOrLoadInnovation(route.params.innovationId, this.authenticationStore.getUserContextInfo())
       .pipe(
-        map(response => {
-          return { id: response.id, name: response.name };
-        }),
+        map(response => ({ id: response.id, name: response.name })),
         catchError(error => {
-          this.contextStore.clearInnovation();
+          this.innovationStoreService.clear$.next();
           this.router.navigateByUrl('error/forbidden-innovation');
-
           this.logger.error('Error fetching data innovation data', error);
           return of(null);
         })
