+++ src/modules/stores/context/context.store.ts
@@ -4,19 +4,16 @@ import { NGXLogger } from 'ngx-logger';
 import { Observable, of } from 'rxjs';
 import { distinctUntilChanged, map, tap } from 'rxjs/operators';
 
-import { InnovationStatusEnum } from '../innovation/innovation.enums';
 import { Store } from '../store.class';
 import { ContextModel } from './context.models';
 import { ContextService } from './context.service';
 import {
   ContextAssessmentType,
-  ContextInnovationType,
   ContextPageLayoutType,
   ContextPageStatusType,
   ContextSchemaType
 } from './context.types';
 
-import { AuthenticationModel } from '../authentication/authentication.models';
 import { InnovationRecordSchemaInfoType } from '../innovation/innovation-record/innovation-record-schema/innovation-record-schema.models';
 import { InnovationRecordSchemaService } from '../innovation/innovation-record/innovation-record-schema/innovation-record-schema.service';
 import { NotificationCategoryTypeEnum, NotificationContextDetailEnum } from './context.enums';
@@ -48,9 +45,6 @@ export class ContextStore extends Store<ContextModel> {
       map(item => item.status),
       distinctUntilChanged()
     );
-  }
-  innovation$(): Observable<null | ContextInnovationType> {
-    return this.state$.pipe(map(item => item.innovation));
   }
   notifications$(): Observable<ContextModel['notifications']> {
     return this.state$.pipe(map(item => item.notifications));
@@ -92,18 +86,6 @@ export class ContextStore extends Store<ContextModel> {
       next: () => this.updateUserUnreadNotifications(),
       error: error => this.logger.error('Error dismissing user notification', error)
     });
-  }
-
-  updateInnovationNotifications(): void {
-    if (this.state.innovation?.id) {
-      this.contextService.getInnovationNotifications(this.state.innovation.id).subscribe({
-        next: response => {
-          this.state.innovation!.notifications = response.data;
-          this.setState();
-        },
-        error: error => this.logger.error('Error obtaining innovation notifications', error)
-      });
-    }
   }
 
   // Page layout methods.
@@ -174,73 +156,6 @@ export class ContextStore extends Store<ContextModel> {
       return of(this.state.irSchema);
     }
     return this.innovationRecordSchemaService.getLatestSchema().pipe(tap(schema => this.setIrSchema(schema)));
-  }
-
-  // Innovation methods.
-  getInnovation(): ContextInnovationType {
-    if (!this.state.innovation) {
-      console.error('Context has NO innovation');
-      return {
-        id: '',
-        name: '',
-        status: InnovationStatusEnum.CREATED,
-        statusUpdatedAt: null,
-        hasBeenAssessed: false,
-        loggedUser: { isOwner: false },
-        categories: [],
-        otherCategoryDescription: '',
-        countryName: '',
-        description: '',
-        postCode: '',
-        expiryAt: 0
-      };
-    }
-    return this.state.innovation;
-  }
-  setInnovation(data: ContextInnovationType): void {
-    this.state.innovation = data;
-    this.setState();
-  }
-  clearInnovation(): void {
-    this.state.innovation = null;
-    this.setState();
-  }
-
-  updateInnovation(data: Partial<ContextInnovationType>): void {
-    if (!this.state.innovation) {
-      console.error('Context has NO innovation');
-      return;
-    }
-
-    if (data.name) {
-      this.state.innovation.name = data.name;
-    }
-    if (data.status) {
-      this.state.innovation.status = data.status;
-    }
-    if (data.assessment) {
-      this.state.innovation.assessment = data.assessment;
-    }
-
-    this.setState();
-  }
-
-  getOrLoadInnovation(
-    innovationId: string,
-    context: AuthenticationModel['userContext']
-  ): Observable<ContextInnovationType> {
-    if (
-      this.state.innovation &&
-      this.state.innovation.id === innovationId &&
-      Date.now() < this.state.innovation.expiryAt
-    ) {
-      return of(this.state.innovation);
-    }
-    return this.contextService.getInnovationContextInfo(innovationId, context).pipe(
-      tap(innovation => {
-        this.setInnovation(innovation);
-      })
-    );
   }
 
   // Assessment methods
