--- src/modules/stores/context/index.ts
@@ -1,2 +1,0 @@
-// Types.
-export { ContextInnovationType } from './context.types';
