+++ src/modules/stores/index.ts
@@ -8,7 +8,6 @@ export { AuthenticationService } from './authentication/authentication.service';
 // // Context Store.
 export { ContextStore } from './context/context.store';
 export { ContextService } from './context/context.service';
-export { ContextInnovationType } from './context/context.types';
 
 // // Innovation Store.
 export { InnovationStore } from './innovation/innovation.store';
@@ -17,3 +16,9 @@ export { InnovationService } from './innovation/innovation.service';
 // // IR Schema Store.
 export { InnovationRecordSchemaStore } from './innovation/innovation-record/innovation-record-schema/innovation-record-schema.store';
 export { InnovationRecordSchemaService } from './innovation/innovation-record/innovation-record-schema/innovation-record-schema.service';
+
+// Ctx
+// // Innovation
+export { InnovationContextService } from './ctx/innovation/innovation-context.service';
+export { InnovationContextStore } from './ctx/innovation/innovation-context.store';
+export { ContextInnovationType } from './ctx/innovation/innovation-context.types';
