+++ src/modules/stores/stores.module.ts
@@ -10,6 +10,8 @@ import { InnovationStore } from './innovation/innovation.store';
 import { InnovationService } from './innovation/innovation.service';
 import { InnovationRecordSchemaStore } from './innovation/innovation-record/innovation-record-schema/innovation-record-schema.store';
 import { InnovationRecordSchemaService } from './innovation/innovation-record/innovation-record-schema/innovation-record-schema.service';
+import { InnovationContextStore } from './ctx/innovation/innovation-context.store';
+import { InnovationContextService } from './ctx/innovation/innovation-context.service';
 
 @NgModule({
   providers: [
@@ -21,6 +23,9 @@ import { InnovationRecordSchemaService } from './innovation/innovation-record/in
 
     InnovationStore,
     InnovationService,
+
+    InnovationContextStore,
+    InnovationContextService,
 
     InnovationRecordSchemaStore,
     InnovationRecordSchemaService
